\name{breast.chr17}
\docType{data}
\alias{breast.chr17}
\title{Breast cancer data set DNA copy number and mRNA expression
  measure on chromosome 17}
\description{
 This is a subset of the breast cancer data from Pollack et al. (2002).
}
\usage{data(breast.chr17)}
\format{
  A list object containing two data matrices: DNA and mRNA. They consist
  chromosome 17 data in 41 samples (4 cell lines and 37 primary tumors).} 
\source{This data can be downloaded at http://www.pnas.org/content/99/20/12963/suppl/DC1}
\references{Pollack, J.R. et al. (2002) Microarray analysis reveals a
  major direct role of DNA copy number alteration in the transcriptional
  program of human breast tumors. Proc. Natl Acad. Sci.  USA, 99,
  12963-12968. }
\keyword{datasets}
