\name{coord}
\docType{data}
\alias{coord}
\title{genomic coordinates}
\description{
 genomic coordinates for the copy number data in gbm
}
\usage{data(coord)}
\format{
 A data matrix consists of chr number, start and end position for the genes included in the gbm copy number data.} 
\references{
Ronglai Shen, Qianxing Mo, Nikolaus Schultz, Venkatraman E. Seshan, Adam B. Olshen, Jason Huse, Marc Ladanyi, Chris Sander. (2012). Integrative Subtype Discovery in Glioblastoma Using iCluster. \emph{PLoS ONE} 7, e35236 

}
\keyword{datasets}
