\name{glp}
\docType{data}
\alias{glp}
\title{good lattice points using the uniform design}
\description{
 good lattice points using the uniform design (Fang and Wang 1995)
}
\usage{data(glp)}
\format{
  A list object containing sampling design for s=2-5 where s is the number of tuning parameters.} 
\references{
Ronglai Shen, Qianxing Mo, Nikolaus Schultz, Venkatraman E. Seshan, Adam B. Olshen, Jason Huse, Marc Ladanyi, Chris Sander. (2012). Integrative Subtype Discovery in Glioblastoma Using iCluster. \emph{PLoS ONE} 7, e35236 

Fang K, Wang Y (1994) Number theoretic methods in statistics. London, UK: Chapman abd Hall.
}
\keyword{datasets}
