\name{simu.datasets}
\docType{data}
\alias{simu.datasets}
\title{simulated dataset. 
}
\description{
Simulated dataset consists of n=150 samples that fall into three clusters and a total of 200 feature.
}
\usage{data(simu.datasets)}
\format{A list object of two data matrices each is of dimension 150 by 200. 
 } 

\keyword{datasets}
